split DocumentEditor/code-block code into shared/ui
