+++ packages/fields-document/src/DocumentEditor/index.tsx
@@ -36,7 +36,7 @@ import { getPlaceholderTextForPropPath } from './component-blocks/utils'
 import { withBlockquote } from './blockquote-shared'
 import { type Relationships, withRelationship } from './relationship'
 import { withDivider } from './divider'
-import { withCodeBlock } from './code-block'
+import { withCodeBlock } from './code-block-shared'
 import { withMarks } from './marks'
 import { renderLeaf } from './leaf'
 import { withSoftBreaks } from './soft-breaks'
